gpu install script

+++ Tuning/gpu_install.pl
@@ -1,14 +1,47 @@
+=b
+  One or more modprobe configuration files to disable Nouveau have been written.  For some distributions, this may be sufficient to disable Nouveau; other       
+  distributions may require modification of the initial ramdisk.  Please reboot your system and attempt NVIDIA driver installation again.  Note if you later     
+  wish to re-enable Nouveau, you will need to delete these files: /usr/lib/modprobe.d/nvidia-installer-disable-nouveau.conf,
+  /etc/modprobe.d/nvidia-installer-disable-nouveau.conf
+
+  WARNING: One or more modprobe configuration files to disable Nouveau are already present at: /usr/lib/modprobe.d/nvidia-installer-disable-nouveau.conf,        
+           /etc/modprobe.d/nvidia-installer-disable-nouveau.conf.  Please be sure you have rebooted your system since these files were written.  If you have     
+           rebooted, then Nouveau may be enabled for other reasons, such as being included in the system initial ramdisk or in your X configuration file.        
+           Please consult the NVIDIA driver README and your Linux distribution's documentation for details on how to correctly disable the Nouveau kernel        
+           driver.
+
+           ERROR: Installation has failed.  Please see the file '/var/log/nvidia-installer.log' for details.  You may find suggestions on fixing installation problems in 
+         the README available on the Linux driver download page at www.nvidia.com.
+
+RTX2060 download link:
+https://us.download.nvidia.com/XFree86/Linux-x86_64/470.74/NVIDIA-Linux-x86_64-470.74.run
+
+=cut
 use Parallel::ForkManager;
 use Cwd;
 #my $currentPath = getcwd();
-$forkNo = 1;
+
+$forkNo = 2;
 my $pm = Parallel::ForkManager->new("$forkNo");
-my $output = "./gpu_cardInfo.dat";
-`rm -f $output`;
-`touch $output`;
+my @nodes = (40..41);
 
-my @nodes = (39-42);
+#+++++++++++ parameters you need to assign correctly!!!!!
+my $gpu_info = "no";#check and output gpu card information for all nodes
+#!!! if yes for $blacklist4nouveau, you need to set @nodes for gpu nodes only
+my $blacklist4nouveau = "no";#make /etc/modprobe.d/blacklist-nouveau.conf or not
+#after rebooting, use the following to install gpu card driver
+#use a larger value for $forkNo 
+my $install_driver = "no";#instll nvidai driver(not work currently)
+my $driver = "/home/rtx2060/NVIDIA-Linux-x86_64-470.74.run";
+#++++++++++++++++++++++
 
+$output;#gpu_info output file
+if($gpu_info eq "yes"){
+    $output = "./gpu_cardInfo.dat";
+    `rm -f $output`;
+    `touch $output`;
+}
+ 
 for (@nodes){
 
 $pm->start and next;
@@ -17,80 +50,60 @@ $pm->start and next;
     $cmd = "ssh $nodename ";
     print "****Check $nodename status\n ";
     #`echo "***$nodename" >> $output`;
-#ping test
-    system("ping -c 1 $nodename");
-    
-    if($? ne 0){
-        `echo "" >> $output`;
-        `echo "??????ping failed at $nodename" >> $output`;
-    }
-    else{#ping ok
-        `echo "" >> $output`;        
-        `echo "******ping ok at $nodename" >> $output`;  
-    #slurmd check
-    my @slurmd = `$cmd 'systemctl status slurmd|egrep "inactive|failed"'`;
-    my $sinfo = `sinfo -R|grep $nodename`;# unexpectedly reboot, slurmd could be active, but resume is still needed. 
-    print "@slurmd\n";
-    if(@slurmd or $sinfo){
-        `echo "???slurmd is inactive or failed at $nodename" >> $output`;
-        `echo "***doing restart slurmd at $nodename" >> $output`;        
-        `$cmd 'systemctl restart slurmd'`;
 
-        #check again
-        @slurmd = `$cmd 'systemctl status slurmd|egrep "inactive|failed"'`;        
-        system("scontrol update nodename=$nodename state=resume");
-        if(@slurmd){`echo "???***slurmd still failed at $nodename after restart slurmd!!!!" >> $output`;}
-    }
-    else{
-        `echo "slurmd is active at $nodename" >> $output`;
-    }
-    #chomp $nodename;
-    #unless($?){system("$cmd 'systemctl restart slurmd'");}
-#
+#get Nvidia GPU card information for finding the proper driver
+    if($gpu_info eq "yes"){
+        my @gpucard = `$cmd 'lspci|grep NVIDIA|grep VGA'`;
+        chomp @gpucard;
+        my $allGPU = join(" ",@gpucard);
+        if(@gpucard){
+            `echo "" >> $output`;
+            `echo "$nodename GPU card info: $allGPU" >> $output`;
+            my @lsmod = `$cmd "lsmod | grep nouveau"`;
+            chomp @lsmod;
+            my $lsmod = join(" ",@lsmod);
+            `echo "lsmod output $lsmod" >> $output`;            
+        }
+        else{#no GPU card
+            `echo '' >> $output`;        
+            `echo "No GPU card in $nodename" >> $output`;  
+        }# 
+    }#gpu_card information
 
-##scp test and remote cp test  
-#    system("scp -o ConnectTimeout=10 scptest.dat root\@$nodename:/root");
-#    if($?){`echo "scp failed at $nodename" >> $output`;}
-#    system("$cmd 'cp scptest.dat /root/");
-#    if($?){`echo "cp to root folder failed at $nodename" >> $output`;}
-#
-##nfs test
-    my @mount = `$cmd 'mount|grep nfs'`;
-    chomp @mount;
-    my @nfs = grep (($_=~m{master:/home|master:/opt}),@mount) ;
-    unless(@nfs){
-        `echo "nfs failed at $nodename" >> $output`;
-        `echo "doing mount -a at $nodename" >> $output`;        
-        `$cmd 'mount -a'`;
-        #check again
-        @mount = `$cmd 'mount|grep nfs'`;
-        @nfs = grep (($_=~m{master:/home|master:/opt}),@mount);        
-        unless(@nfs){`echo "???***nfs still failed at $nodename after mount -a!!!!" >> $output`;}
+#set blacklist for nouveau 
+    if($blacklist4nouveau eq "yes"){
+        my $file = `$cmd "ls /etc/modprobe.d/blacklist-nouveau.conf"`;
+        chomp $file;
+        if($?){#no blacklist-nouveau.conf
+            `$cmd 'touch /etc/modprobe.d/blacklist-nouveau.conf'`;
+            `$cmd 'echo "blacklist nouveau" >> /etc/modprobe.d/blacklist-nouveau.conf'`;
+            `$cmd 'echo "options nouveau modeset=0" >> /etc/modprobe.d/blacklist-nouveau.conf'`;
+            `$cmd 'dracut --force'`;
+            `$cmd 'reboot'`;
+        }
+        else{#blacklist-nouveau.conf exists!!
+            my $grepKeyW = `$cmd "grep 'blacklist' /etc/modprobe.d/blacklist-nouveau.conf"`;
+            unless($grepKeyW){
+                `$cmd 'echo "blacklist nouveau" >> /etc/modprobe.d/blacklist-nouveau.conf'`;
+            }
+            $grepKeyW = `$cmd "grep 'options nouveau' /etc/modprobe.d/blacklist-nouveau.conf"`;
+            unless($grepKeyW){
+                `$cmd 'echo "options nouveau modeset=0" >> /etc/modprobe.d/blacklist-nouveau.conf'`;
+            }
+        }
         
+    }#blacklist4nouveau
+
+    if($install_driver eq "yes"){
+       system("$cmd \"echo -e 'Accept'|$driver\"");
+       print "test nvidia-smi for $nodename\n";
+       `$cmd "nvidia-smi"`;
+
     }
-    else{
-        `echo "nfs good at $nodename" >> $output`;
-    }
-##munge test
-#    system("munge -n \| ssh $nodename unmunge");
-#    if($?){`echo "munge failed at $nodename" >> $output`;} 
 
-#swap test
 
-    #system ("$cmd 'free'");
-    my $swap = `$cmd 'free|grep Swap:|awk "{print \\\$2}"'`;
-    chomp $swap;
-    unless($swap){
-        `$cmd 'rm -f /swap/*'`;
-        `$cmd 'dd if=/dev/zero of=/swap/swap bs=1M count=4096'`;
-        system("$cmd 'chmod 0644 /swap/swap'");
-        `$cmd 'mkswap -f /swap/swap'`;
-        `$cmd 'swapon /swap/swap'`;
-        system("$cmd 'swapon -s'");
-    }#swap
-
-    }# good ping loop 
    $pm->finish;
+
 }
 $pm->wait_all_children;
